backend/requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Google ADK and AI (REQUIRED for capstone)
google-adk>=0.1.0
google-generativeai>=0.3.0

# Database
pymongo>=4.5.0
motor>=3.3.0

# Firebase Admin
firebase-admin>=6.2.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# CORS and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Rate Limiting (optional)
slowapi>=0.1.8

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
